refactor: move HeaderNavigation out of the Spaces feature to common components
