+++ apps/web/src/components/common/Header/Topbar/index.tsx
@@ -4,7 +4,7 @@ import { Menu } from 'lucide-react'
 import { usePathname } from 'next/navigation'
 import { Button } from '@/components/ui/button'
 import { AppRoutes } from '@/config/routes'
-import { HeaderNavigation } from '@/features/spaces/components/HeaderNavigation'
+import { HeaderNavigation } from '@/components/common/Header/HeaderNavigation'
 import { useLoadFeature } from '@/features/__core__'
 import { WalletFeature, useWalletPopover } from '@/features/wallet'
 import { GlobalSearchFeature } from '@/features/global-search'
